requirement.txt
absl-py==1.0.0
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
atari-py==0.2.6
attrs==21.2.0
audioread==2.1.9
bchlib==0.14.0
boltons==21.0.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
comet-ml==3.22.1
configobj==5.0.6
configparser==5.1.0
cycler==0.11.0
Cython==0.29.24
decorator==5.1.0
docker-pycreds==0.4.0
dulwich==0.20.26
everett==2.0.1
faiss==1.7.2
fire==0.5.0
flatbuffers==2.0
gast==0.2.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.6.0
idna==3.3
imageio==2.21.1
importlib-metadata==4.8.2
importlib-resources==5.4.0
joblib==1.1.0
jsonschema==4.2.1
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
librosa==0.8.1
llvmlite==0.37.0
loguru==0.5.3
Markdown==3.3.5
matplotlib==3.4.3
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.3
munch==2.5.0
networkx==2.6.3
numba==0.54.1
numpy==1.20.3
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opencv-python==4.5.4.58
openTSNE==0.6.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.0
pathtools==0.1.2
Pillow==8.4.0
pip==21.2.4
plotly==5.3.1
pooch==1.5.2
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
pyworld==0.3.0
PyYAML==5.3.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resampy==0.2.2
Resemblyzer==0.1.1.dev0
rsa==4.7.2
scikit-image==0.19.3
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.12.2
semantic-version==2.8.5
sentry-sdk==1.5.0
setproctitle==1.2.3
setuptools==61.2.0
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
SoundFile==0.10.3.post1
subprocess32==3.5.4
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
torch==1.9.0+cu111
torchvision==0.10.0
tqdm==4.48.2
typing==3.7.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wandb==0.12.14
webrtcvad==2.0.10
websocket-client==1.2.1
Werkzeug==2.0.2
wheel==0.37.0
win32-setctime==1.0.3
wincertstore==0.2
wrapt==1.13.3
wurlitzer==3.0.2
yacs==0.1.8
yaspin==2.1.0
zipp==3.6.0
